requirements.txt
fastapi
uvicorn
pypdf
reportlab
google-genai
python-multipart
jinja2